Add C++ implementations of the GaussianBlurEffect class. (#398)
* Add C++ implementations of the GaussianBlurEffect class.

Effect classes are defined in Win2D, but in C# we avoid taking the Win2D dependency by implementing the class in the generated code.

This change also cleans up the CompositeEffect class to be a bit clearer and more consistent with GaussianBlurEffect. And fixes a compilation error in the CX output.