Add: version checking for Blur effect, path information for error message.
Blur effects will only work on UAP version 8 and above. Previously if you tried to run LottieGen to generate code that could run on version 7 and the Lottie file used blur, we'd throw an exception. Now we report the issue and turn off the blur.

Designers have asked for more context for the LT0035 error message, so I've added information about the layer and ancestor precomp layers. This information could be added to other issue messages in future.