Convert LottieGen to .NET 5 and enable LottieGen.win-x64
With this change, we'll produce a new nuget package that contains the LottieGen.exe executable built into a single file with no dependencies (not even .NET).
This is being done to enable LottieGen to be more easily used in a build. The nuget package contains the .exe that will run on any x64 machine, and a props file that can be used by msbuild targets to allow them to find the .exe.