Fix some missing simplifications in Composition expressions. (#429)
* Fix some missing simplifications in Composition expressions.

The logic for determining whether an expression could be simplified was sometimes wrong because it ended up hitting an overridden equality operator which compared the already-simplified values.

Also fixed the test script to look at last write time when determining which is the latest build of LottieGen.exe to run. We were looking at creation time, but it turns out that when the compiler builds, it doesn't cause the creation time to be updated if the output file previously existed. Our test script could accidentally use a release build when the latest build was actually a debug build, resulting in the test running on an old build.

* Replace the IsAtomic bool with a more generalized Precedence concept.

Without this, it was getting unwieldy to handle all the cases of precedence. Now we're removing parentheses more aggressively, so the expressions will be more efficient to parse.

* Eliminate some unnecessary whitespace in Min and Max expression functions.