LottieViewer UI improvements for AnimatedIcon (#426)
* Tweak the nudging factor for markers, and better UI for markers.

These changes are to improve AnimatedIcon workflow.
The nudging value of 0.01 was insufficient, so this ups it to 0.05.
The UI in LottieViewer for markers and the scrubber is confusing because it shows progress, which designers don't care about. So now it uses frames everywhere. Developers that need to convert to progress can do the math themselves (it's just (frame count) / (number of frames) plus a nudge.

* Make the arrow keys on the scrubber move by a single frame.

* LottieViewer UI now only deals in frames.

Seeing as LottieViewer is for designers, and they don't know about progress, the markers and scrubber will now only show frames information.
Also added more frame information to the Lottie properties.
Arrow keys now move by a single frame on each press.

* Fix the colors on hyperlinks and the playback rate slider.

The default colors for hyperlinks and sliders are themed, which can cause the colors to become unreadble on the gray background of the main page..