Fix for #406 (#433)
When using the "-AdditionalInterface" parameter in LottieGen (which causes the generated code to declare that it implements interfaces that it otherwise doesn't know about), the interface needs to be known to the generated IDL file.
There are 2 ways that the IDL may know about the interface:
1. The interface may exist in a .winmd file that is referenced by the project.
2. The interface may be defined in the project using an IDL file.

LottieGen doesn't know which case the user requires. Previously we were assuming case 2 which causes problems if the interface is not in an IDL file.
Rather than requiring the user to tell LottieGen which case is being used, we know wrap the import in a condition that depends on the existence of the file. This should handle both cases seamlessly.